+++ app/templates/components/github-label.hbs
@@ -1,1 +1,1 @@
-{{labelWithEmoji}}
+{{label-with-emoji label}}
